panelcleaner: An interactive interface to homogenize messy panel data into a long format

Panel data structure can sometimes change over the course of data collection, which can be a real nuisance when cleaning data. This package aims to document the state of all waves of data in a comprehensive manner and then homogenize the data into a long dataset ready for data transformations.

Details

The panelcleaner package addresses data structure issues with longitudinal data by defining two data structures: a tk_panel class that is a wrapper around the waves of tabular data, and a panel_map class which is a type of data.frame that outlines the raw data content (variable names and categorical levels) and a common "homogenized" form that is consistent across all waves.

Resources for workflow:

Defining a panel from waves of data: enpanel(), add_wave()

Accessing waves of data from an unhomogenized panel: wave()

Defining a panel mapping: panel_mapping()

Adding a panel mapping to a panel: add_mapping()

Homogenize a panel: homogenize_panel()

Combine waves and extract long-form data.frame: bind_waves()

See also

Author

Maintainer: Patrick Anker psanker@nyu.edu (ORCID)

Other contributors:

Global TIES for Children (https://research.steinhardt.nyu.edu/ihdsc/global-ties) [copyright holder]